+++ src/oci/monitoring/models/metric.py
@@ -28,6 +28,10 @@ class Metric(object):
             The value to assign to the namespace property of this Metric.
         :type namespace: str
 
+        :param resource_group:
+            The value to assign to the resource_group property of this Metric.
+        :type resource_group: str
+
         :param compartment_id:
             The value to assign to the compartment_id property of this Metric.
         :type compartment_id: str
@@ -40,6 +44,7 @@ class Metric(object):
         self.swagger_types = {
             'name': 'str',
             'namespace': 'str',
+            'resource_group': 'str',
             'compartment_id': 'str',
             'dimensions': 'dict(str, str)'
         }
@@ -47,12 +52,14 @@ class Metric(object):
         self.attribute_map = {
             'name': 'name',
             'namespace': 'namespace',
+            'resource_group': 'resourceGroup',
             'compartment_id': 'compartmentId',
             'dimensions': 'dimensions'
         }
 
         self._name = None
         self._namespace = None
+        self._resource_group = None
         self._compartment_id = None
         self._dimensions = None
 
@@ -111,6 +118,38 @@ class Metric(object):
         :type: str
         """
         self._namespace = namespace
+
+    @property
+    def resource_group(self):
+        """
+        Gets the resource_group of this Metric.
+        Resource group provided with the posted metric. A resource group is a custom string that can be used as a filter. Only one resource group can be applied per metric.
+        A valid resourceGroup value starts with an alphabetical character and includes only alphanumeric characters, periods (.), underscores (_), hyphens (-), and dollar signs ($).
+        Avoid entering confidential information.
+
+        Example: `frontend-fleet`
+
+
+        :return: The resource_group of this Metric.
+        :rtype: str
+        """
+        return self._resource_group
+
+    @resource_group.setter
+    def resource_group(self, resource_group):
+        """
+        Sets the resource_group of this Metric.
+        Resource group provided with the posted metric. A resource group is a custom string that can be used as a filter. Only one resource group can be applied per metric.
+        A valid resourceGroup value starts with an alphabetical character and includes only alphanumeric characters, periods (.), underscores (_), hyphens (-), and dollar signs ($).
+        Avoid entering confidential information.
+
+        Example: `frontend-fleet`
+
+
+        :param resource_group: The resource_group of this Metric.
+        :type: str
+        """
+        self._resource_group = resource_group
 
     @property
     def compartment_id(self):